Advanced: - write pedal table only in settings - read and display pedal table on load
git-svn-id: svn://chopstick.local/softstep_editor@313 b087e19c-e0c3-4903-a6fc-ecf072d2c2be

+++ standalone/advanced/tableinterface.h
@@ -30,7 +30,7 @@ signals:
 
 
 public slots:
-    void slotDrawTable(float x, float y, int width);
+    void slotDrawTable(float x, float y, float width);
     void slotClearTable();
     void slotDrawLinear();
